Tidy up assertion code
